运动会多人报名项目后端

+++ scms/src/main/java/com/ashuo/scms/entity/Record.java
@@ -2,6 +2,8 @@ package com.ashuo.scms.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.io.Serializable;
 
@@ -32,7 +34,7 @@ public class Record implements Serializable {
 
     private User user;
 
-    private Integer recordScore;
+    private BigDecimal recordScore;
 
     private String recordStatus;
